requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
celery
redis
python-dotenv
langchain
langchain-google-genai
google-api-core


gevent
langchain-openai 
langchain-core
python-dotenv
openai
bcrypt
python-jose[cryptography]
python-multipart
alembic
pydantic-settings
PyYAML>=6.0.1
pytest
httpx